refactor(check): extract state.ts from checker.ts
Move state management types and functions to dedicated state.ts module:
- ExprResult, UnreachableRange, MatchContext interfaces
- BlockContext types (TypeDeclContext, RecordLiteralContext, NestedRecordInitContext)
- CheckerState interface
- Block context stack functions (push/pop/current/parent)
- Context checking functions (isInTypeDeclContext, etc.)

This is layer 0 of the checker modularization, with no dependencies on
other checker modules.
