+++ packages/compiler/src/check/checker.ts
@@ -27,117 +27,25 @@ import {
 	type FieldInfo,
 	type InstId,
 	InstKind,
-	type Scope,
 	type SymbolId,
 	type TypeId,
 } from './types.ts'
-
-interface UnreachableRange {
-	firstNodeId: NodeId
-	startLine: number
-	endLine: number
-}
-
-/**
- * Context for collecting match arms.
- */
-interface MatchContext {
-	/** Scrutinee expression result */
-	scrutinee: ExprResult
-	/** Scrutinee node ID (for diagnostics) */
-	scrutineeNodeId: NodeId
-	/** Expected result type of the match */
-	expectedType: TypeId
-	/** Collected arms (pattern + body) */
-	arms: Array<{ patternNodeId: NodeId; bodyInstId: InstId }>
-	/** The node ID of the match binding/expr for diagnostics */
-	matchNodeId: NodeId
-	/** Binding name for variable creation after finalization */
-	bindingNameId: StringId
-	/** Binding node ID for symbol creation */
-	bindingNodeId: NodeId
-}
-
-/**
- * Common properties shared by all block context kinds.
- */
-interface BlockContextBase {
-	typeName: string
-	typeId: TypeId | null
-	nodeId: NodeId
-	children: NodeId[]
-	expectedChildKind: NodeKind
-	childIndentLevel?: number
-	fieldNames: Set<string>
-}
-
-/**
- * Context for type declaration blocks.
- */
-interface TypeDeclContext extends BlockContextBase {
-	kind: 'TypeDecl'
-	fields: Array<{ name: string; typeId: TypeId; nodeId: NodeId }>
-}
-
-/**
- * Context for record literal blocks (top-level record instantiation).
- */
-interface RecordLiteralContext extends BlockContextBase {
-	kind: 'RecordLiteral'
-	bindingNameId: StringId
-	bindingNodeId: NodeId
-	fieldInits: Array<{ name: string; nodeId: NodeId; exprResult: ExprResult }>
-}
-
-/**
- * Context for nested record initialization blocks.
- */
-interface NestedRecordInitContext extends BlockContextBase {
-	kind: 'NestedRecordInit'
-	fieldName: string
-	parentPath: string
-	fieldInits: Array<{ name: string; nodeId: NodeId; exprResult: ExprResult }>
-}
-
-/**
- * Discriminated union for block-based constructs.
- * All share the pattern: header starts block, indented lines are children, dedent finalizes.
- */
-type BlockContext = TypeDeclContext | RecordLiteralContext | NestedRecordInitContext
-
-function pushBlockContext(state: CheckerState, ctx: BlockContext): void {
-	state.blockContextStack.push(ctx)
-}
-
-function popBlockContext(state: CheckerState): BlockContext | null {
-	return state.blockContextStack.pop() ?? null
-}
-
-function currentBlockContext(state: CheckerState): BlockContext | null {
-	return state.blockContextStack.at(-1) ?? null
-}
-
-function parentBlockContext(state: CheckerState): BlockContext | null {
-	return state.blockContextStack.at(-2) ?? null
-}
-
-void parentBlockContext
-
-interface CheckerState {
-	readonly insts: InstStore
-	readonly scopes: ScopeStore
-	readonly symbols: SymbolStore
-	readonly types: TypeStore
-	currentScope: Scope
-	unreachableRange: UnreachableRange | null
-	matchContext: MatchContext | null
-	blockContextStack: BlockContext[]
-}
-
-interface ExprResult {
-	typeId: TypeId
-	instId: InstId | null
-}
+import {
+	type BlockContext,
+	type CheckerState,
+	type ExprResult,
+	type MatchContext,
+	type NestedRecordInitContext,
+	type RecordLiteralContext,
+	type TypeDeclContext,
+	currentBlockContext,
+	isInNestedRecordInitContext,
+	isInRecordInitContext,
+	isInRecordLiteralContext,
+	isInTypeDeclContext,
+	popBlockContext,
+	pushBlockContext,
+} from './state.ts'
 
 function isValidExprResult(result: ExprResult): result is { typeId: TypeId; instId: InstId } {
 	return result.typeId !== BuiltinTypeId.Invalid && result.instId !== null
@@ -2535,14 +2443,6 @@ function finalizeRecordLiteral(state: CheckerState, context: CompilationContext)
 }
 
 /**
- * Check if we're in a NestedRecordInit block context.
- */
-function isInNestedRecordInitContext(state: CheckerState): boolean {
-	const ctx = currentBlockContext(state)
-	return ctx !== null && ctx.kind === 'NestedRecordInit'
-}
-
-/**
  * Validate that the nested record init type matches the expected field type.
  * Returns true if valid, false if there's a mismatch (error already emitted).
  */
@@ -3101,22 +3001,6 @@ function processIndentedLineAsMatchArm(
 	return true
 }
 
-/**
- * Check if we're in a TypeDecl block context.
- */
-function isInTypeDeclContext(state: CheckerState): boolean {
-	const ctx = currentBlockContext(state)
-	return ctx?.kind === 'TypeDecl'
-}
-
-/**
- * Check if we're in a RecordLiteral block context.
- */
-function isInRecordLiteralContext(state: CheckerState): boolean {
-	const ctx = currentBlockContext(state)
-	return ctx !== null && ctx.kind === 'RecordLiteral'
-}
-
 function processIndentedLineAsFieldDecl(
 	lineId: NodeId,
 	state: CheckerState,
@@ -3140,14 +3024,6 @@ function processIndentedLineAsFieldDecl(
 	}
 
 	return false
-}
-
-/**
- * Check if we're in a RecordLiteral or NestedRecordInit block context.
- */
-function isInRecordInitContext(state: CheckerState): boolean {
-	const ctx = currentBlockContext(state)
-	return ctx !== null && (ctx.kind === 'RecordLiteral' || ctx.kind === 'NestedRecordInit')
 }
 
 /**
